requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
statsmodels
xgboost
lightgbm
optuna
shap
joblib
flask
tabulate
pytest
polars
pyarrow
pymongo
pydantic
PyJWT
bcrypt
flask-cors
